separated engine and game source code, added filters to solution
