+++ Raycasting/YSElog.txt
@@ -539,3 +539,48 @@ Log started: 7 Dec 2020 5:58:35pm
 === start of YSE log ===
 Log started: 7 Dec 2020 6:05:15pm
 
+
+**********************************************************
+=== start of YSE log ===
+Log started: 7 Dec 2020 7:53:51pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 7 Dec 2020 7:54:47pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 8:34:05pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 8:49:30pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 9:00:44pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 9:08:48pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 9:09:12pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 9:10:14pm
+
+
+**********************************************************
+=== start of YSE log ===
+Log started: 9 Dec 2020 9:15:55pm
+
