+++ Raycasting/src/engine/Geometry.cpp
@@ -2,7 +2,7 @@
 #include "Geometry.h"
 
 #include <stdio.h>
-#include "Window.h"
+#include "engine/Window.h"
 
 namespace {
 	bool getRectRectIntersection(const Geo::Rectangle& rect1, const Geo::Rectangle& rect2, std::vector<std::array<float, 2>>* poi = nullptr) {
