+++ Raycasting/src/engine/SeeableLine.cpp
@@ -1,4 +1,4 @@
-#include "SeeableLine.h"
+#include "engine/SeeableLine.h"
 
 SeeableLine::SeeableLine(float x1, float y1, float x2, float y2, Texture * tex)
 {
