+++ Raycasting/src/engine/Window.h
@@ -6,7 +6,7 @@
 #include <string>
 #include <array>
 
-#include "Texture.h"
+#include "engine/Texture.h"
 
 class Window {
 public:
